Internal Documentation

Documentation for PrimaryCensored.jl's internal interface.

Contents

Contents

Index

Internal API

Index

PrimaryCensored.PrimaryCensoredDist

PrimaryCensored.WithinIntervalCensored

Base.rand

Distributions.logpdf

PrimaryCensored._check_and_give_ts

Internal API

PrimaryCensored.PrimaryCensoredDist — Type

struct PrimaryCensoredDist{D1<:(Distributions.UnivariateDistribution), D2<:(Distributions.UnivariateDistribution)} <: Distributions.UnivariateDistribution{Distributions.ValueSupport}

Generic wrapper for a primary event censored distribution.

Contructors

PrimaryCensoredDist(; uncensored::D, censoring::D): Constructs a PrimaryCensoredDist distribution with the specified uncensored and censoring distributions.

Examples

using PrimaryCensored, Distributions

uncensored = Normal()
censoring = Uniform(0, 1)

d = primarycensored(uncensored, censoring)

rand(d)

Fields

uncensored::Distributions.UnivariateDistribution: The original (uncensored) distribution.

censoring::Distributions.UnivariateDistribution: The primary event censoring distribution.

PrimaryCensored.WithinIntervalCensored — Type

struct WithinIntervalCensored{D<:(Distributions.UnivariateDistribution), L<:Real, U<:Real} <: Distributions.UnivariateDistribution{Distributions.ValueSupport}

Implement a censoring function.

Takes a UnivariateDistribution and returns a WithinIntervalCensored object but depends only on the cdf method.

Arguments

dist the truncated primarycensored distribution

lower the lower bound of the interval

upper the upper bound of the interval

Returns

WithinIntervalCensored object

Fields

dist::Distributions.UnivariateDistribution

lower::Real

upper::Real

Base.rand — Method

rand(
    rng::Random.AbstractRNG,
    d::PrimaryCensored.WithinIntervalCensored
) -> Any

Generate a random sample from a within-interval censored distribution.

Arguments

rng: Random number generator

d: WithinIntervalCensored distribution object

Returns

A random sample from the within-interval censored distribution

Examples

using PrimaryCensored, Distributions
d = truncated(Normal(5,2), 0, 5)
trunc_d = within_interval_censored(d, 2, 4)
rand(trunc_d)

Distributions.logpdf — Method

logpdf(d::PrimaryCensored.WithinIntervalCensored) -> Any

Returning a within interval censored logpdf

Arguments

d an WithinIntervalCensoredDist object

Returns

discretized log probability mass function

Examples

using PrimaryCensored, Distributions
d = truncated(Normal(5,2), 0, 5)
trunc_d = within_interval_censored(d, 2, 4)
logpdf(trunc_d)
rand(trunc_d)

PrimaryCensored._check_and_give_ts — Method

_check_and_give_ts(
    dist::PrimaryCensored.PrimaryCensoredDist,
    Δd,
    D,
    upper
) -> Any

Internal function to check censored_pmf arguments and return the time steps of the rightmost limits of the censor intervals.