Public Documentation

Documentation for PrimaryCensored.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Contents

Index

Public API

Index

PrimaryCensored.censored_cdf

PrimaryCensored.censored_pmf

PrimaryCensored.primarycensored

PrimaryCensored.within_interval_censored

Public API

PrimaryCensored.censored_cdf — Method

censored_cdf(
    dist::PrimaryCensored.PrimaryCensoredDist;
    Δd,
    D,
    upper
) -> Any

Create a discrete probability cumulative distribution function (CDF) from a given PrimaryCensoredDist, specialized for the case where all observsations are censored into disjoint intervals of width Δd. A canonical example of this would be daily or weekly reporting.

NB: censored_cdf returns the non-truncated CDF, i.e. the CDF without conditioning on the secondary event occuring either before or after some time. Truncation can be applied by conditioning.

Arguments

dist: The PrimaryCensoredDist distribution from which to evaluate the CDF.

Δd: The step size for discretizing the domain. Default is 1.0.

D: The upper bound of the domain. Must be greater than Δd. Default D = nothing

indicates that the coverage of the distribution returned at its upperth percentile, _if the delay started at the beginning of its primary interval rounded to nearest multiple of Δd.

Returns

A vector representing the CDF with 0.0 appended at the beginning.

Raises

AssertionError if the minimum value of dist is negative.

AssertionError if Δd is not positive.

AssertionError if D is shorter than Δd.

AssertionError if D is not a multiple of Δd.

Examples

using PrimaryCensored, Distributions

dist = primarycensored(Exponential(1.0), Uniform(0, 1))

censored_cdf(dist; D = 10) |>
    p -> round.(p, digits=3)

# output
11-element Vector{Float64}:
 0.0
 0.368
 0.767
 0.914
 0.969
 0.988
 0.996
 0.998
 0.999
 1.0
 1.0

PrimaryCensored.censored_pmf — Method

censored_pmf(
    dist::PrimaryCensored.PrimaryCensoredDist;
    Δd,
    D,
    upper
) -> Any

Create a discrete probability probability mass function (PMF) from a given PrimaryCensoredDist, specialized for the case where all observsations are censored into disjoint intervals of width Δd. A canonical example of this would be daily or weekly reporting.

NB: censored_pmf returns the non-truncated PMF, i.e. the PMF without conditioning on the secondary event occuring either before or after some time. Truncation can be applied by conditioning.

Arguments

dist: The PrimaryCensoredDist distribution from which to evaluate the CDF.

Δd: The step size for discretizing the domain. Default is 1.0.

D: The upper bound of the domain. Must be greater than Δd. Default D = nothing

indicates that the coverage of the distribution returned at its upperth percentile, _if the delay started at the beginning of its primary interval rounded to nearest multiple of Δd.

Returns

A vector representing the CDF with 0.0 appended at the beginning.

Raises

AssertionError if the minimum value of dist is negative.

AssertionError if Δd is not positive.

AssertionError if D is shorter than Δd.

AssertionError if D is not a multiple of Δd.

Examples

using PrimaryCensored, Distributions

dist = primarycensored(Exponential(1.0), Uniform(0, 1))

censored_pmf(dist; D = 10) |>
    p -> round.(p, digits=3)

# output
10-element Vector{Float64}:
 0.368
 0.4
 0.147
 0.054
 0.02
 0.007
 0.003
 0.001
 0.0
 0.0

PrimaryCensored.primarycensored — Method

primarycensored(
    uncensored::Distributions.UnivariateDistribution,
    censoring::Distributions.UnivariateDistribution
) -> PrimaryCensored.PrimaryCensoredDist

Constructs a `PrimaryCensoredDist` distribution with the specified uncensored and censoring distributions.

# Arguments
- `uncensored::D`: The uncensored distribution.
- `censoring::D`: The primary censoring distribution.

# Returns
- A `PrimaryCensored` distribution.

# Examples
```@example
using PrimaryCensored, Distributions

uncensored = Normal(4, 1)
censoring = Uniform(0, 1)

d = primarycensored(uncensored, censoring)

rand(d, 10)

x = 0:10
cdf.(d, x)
```

PrimaryCensored.within_interval_censored — Method

within_interval_censored(
    dist::Distributions.UnivariateDistribution,
    lower::Real,
    upper::Real
) -> PrimaryCensored.WithinIntervalCensored

Generic wrapper for within window censored an object

Contructors